requirements.txt
mysql-connector-python==8.0.22
protobuf==3.14.0     
six==1.15.0